requirements.txt
transformers
pytesseract
pyyaml
tesseract-ocr